add: KeyboardToolbarSectionを追加(SettingViewからの切り出し
